% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalBoosting.R
\name{causalBoosting}
\alias{causalBoosting}
\title{Fit a causal boosting model}
\usage{
causalBoosting(x, tx, y, num.trees = 500, maxleaves = 4, eps = 0.01,
  splitSpread = 0.1, x.est = NULL, tx.est = NULL, y.est = NULL,
  propensity = FALSE, stratum = NULL, stratum.est = NULL,
  isConstVar = TRUE)
}
\arguments{
\item{x}{matrix of covariates}

\item{tx}{vector of treatment indicators (0 or 1)}

\item{y}{vector of response values}

\item{num.trees}{number of shallow causal trees to build}

\item{maxleaves}{maximum number of leaves per causal tree}

\item{eps}{learning rate}

\item{splitSpread}{how far apart should the candidate splits be for the
causal trees? (e.g. \code{splitSpread = 0.1}) means we consider 10 quantile
cutpoints as candidates for making split}

\item{x.est}{optional matrix of estimation-set covariates used for honest
re-estimation (ignored if \code{tx.est = NULL} or \code{y.est = NULL})}

\item{tx.est}{optional vector of estimation-set treatment indicators
(ignored if \code{x.est = NULL} or \code{y.est = NULL})}

\item{y.est}{optional vector of estimation-set response values
(ignored if \code{x.est = NULL} or \code{y.est = NULL})}

\item{propensity}{logical: should propensity score stratification be used?}

\item{stratum}{optional vector giving propensity score stratum for each
observation (only used if \code{propensity = TRUE})}

\item{stratum.est}{optional vector giving propensity score stratum for each
estimation-set observation (ignored if \code{x.est = NULL} or
\code{tx.est = NULL} or \code{y.est = NULL})}

\item{isConstVar}{logical: for the causal tree splitting criterion
(T-statistc), should it be assumed that the noise variance is the same in
treatment and control arms?}
}
\value{
an object of class \code{causalBoosting} with attributes:
 \itemize{
   \item CBM: a list storing the intercept, the causal trees and \code{eps}
   \item tauhat: matrix of treatment effects for each patient for each step
   \item G1: estimated-treatment conditional mean for each patient
   \item G0: estimated-control conditional mean for each patient
   \item err.y: training error at each step, in predicting response
   \item num.trees: number of trees specified by function call
 }
}
\description{
Fit a causal boosting model
}
\details{
This function exists primarily to be called by cv.causalBoosting because
 the num.trees parameter generally needs to be tuned via cross-validation.
}
\examples{
# Randomized experiment example

n = 100 # number of training-set patients to simulate
p = 10  # number of features for each training-set patient

# Simulate data
x = matrix(rnorm(n * p), nrow = n, ncol = p) # simulate covariate matrix
tx_effect = x[, 1] + (x[, 2] > 0) # simple heterogeneous treatment effect
tx = rbinom(n, size = 1, p = 0.5) # random treatment assignment
y = rowMeans(x) + tx * tx_effect + rnorm(n, sd = 0.001) # simulate response

# Estimate causal boosting model
fit_cb = causalBoosting(x, tx, y, num.trees = 500)
pred_cb = predict(fit_cb, newx = x, num.trees = 500)

# Visualize results
plot(tx_effect, pred_cb, main = 'Causal boosting',
 xlab = 'True treatment effect', ylab = 'Estimated treatment effect')
abline(0, 1, lty = 2)

}
