% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollinated.ranger.R
\name{pollinated.ranger}
\alias{pollinated.ranger}
\title{Pollinate a fitted ranger random forest model}
\usage{
pollinated.ranger(object, x, y)
}
\arguments{
\item{object}{a fitted \code{ranger} object}

\item{x}{matrix of covariates}

\item{y}{vector of response values}
}
\value{
an object of class \code{pollinated.ranger} which is a \code{ranger}
 object that has been pollinated with the data in (x, y)
}
\description{
Pollinate a fitted ranger random forest model
}
