% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bagged.causalMARS.R
\name{predict.bagged.causalMARS}
\alias{predict.bagged.causalMARS}
\title{Make predictions from a bag of fitted causal MARS models}
\usage{
\method{predict}{bagged.causalMARS}(object, newx, type = c("average", "all"),
  ...)
}
\arguments{
\item{object}{a fitted \code{bagged.causalMARS} object}

\item{newx}{matrix of new covariates for which estimated treatment effects
are desired}

\item{type}{type of prediction required:
'average' returns a vector of the averages of the bootstrap estimates.
'all' returns a matrix of all of the bootstrap estimates.}

\item{...}{ignored}
}
\value{
a vector of estimated personalized treatment effects corresponding
 to the rows of \code{newx}
}
\description{
Make predictions from a bag of fitted causal MARS models
}
