% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.causalMARS.R
\name{predict.causalMARS}
\alias{predict.causalMARS}
\title{Make predictions from a fitted causal MARS model}
\usage{
\method{predict}{causalMARS}(object, newx, active, ...)
}
\arguments{
\item{object}{a fitted \code{causalMARS} object}

\item{newx}{matrix of new covariates for which estimated treatment effects
are desired}

\item{active}{indices of columns with nonzero norm (defaults to model
selected via backward stepwise phase, or the full model if
\code{backstep = FALSE})}

\item{...}{ignored}
}
\value{
a vector of estimated personalized treatment effects corresponding
 to the rows of \code{newx}
}
\description{
Make predictions from a fitted causal MARS model
}
