% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.causalTree.R
\name{predict.causalTree}
\alias{predict.causalTree}
\title{Make predictions from a fitted causal tree model}
\usage{
\method{predict}{causalTree}(object, newx, newtx = NULL,
  type = c("treatment.effect", "conditional.mean", "response"),
  honest = FALSE, naVal = 0, ...)
}
\arguments{
\item{object}{a fitted \code{causalTree} object}

\item{newx}{matrix of new covariates for which estimated treatment effects
are desired}

\item{newtx}{option vector of new treatment assignments
(only used if \code{type = 'response'})}

\item{type}{type of prediction required:
'treatment.effect' returns estimated treatment effect.
'conditional.mean' returns two predictions, one for each arm.
'response' returns prediction for arm corresponding to newtx.}

\item{honest}{logical: should honest re-estimates of leaf means be used for
prediction? This requires that \code{x.est, tx.est, y.est} were specified
when the causal boosting model was fit}

\item{naVal}{value with which to replace \code{NA} predictions}

\item{...}{ignored}
}
\value{
a vector or matrix of predictions corresponding to the rows of
 \code{newx}
}
\description{
Make predictions from a fitted causal tree model
}
