% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pollinated.ranger.R
\name{predict.pollinated.ranger}
\alias{predict.pollinated.ranger}
\title{Make predictions from a pollinated ranger random forest model}
\usage{
\method{predict}{pollinated.ranger}(object, newx, predict.all = FALSE,
  na.treatment = c("omit", "replace", "NA"), ...)
}
\arguments{
\item{object}{a fitted \code{pollinated.ranger} object}

\item{newx}{matrix of new covariates for which predictions are desired}

\item{predict.all}{logical: should predictions from all trees be returned?
Otherwise the average across trees is returned}

\item{na.treatment}{how to treat NA predictions from individual trees:
'omit' only uses trees for which the prediction is not NA.
'replace' replaces NA predictions with the overall mean response.
'NA' returns NA if any tree prediction is NA.}

\item{...}{additional arguments passed on to \code{predict.ranger}}
}
\value{
a vector of predicted treatment effects corresponding to the rows of
 newx
}
\description{
Make predictions from a pollinated ranger random forest model
}
