% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTOforest.R
\name{PTOforest}
\alias{PTOforest}
\title{Fit a pollinated transformed outcome (PTO) forest model}
\usage{
PTOforest(x, tx, y, pscore = rep(0.5, nrow(x)), num.trees = 500,
  mtry = ncol(x), min.node.size = max(25, nrow(x)/40), postprocess = TRUE,
  verbose = FALSE)
}
\arguments{
\item{x}{matrix of covariates}

\item{tx}{vector of treatment indicators (0 or 1)}

\item{y}{vector of response values}

\item{pscore}{vector of propensity scores}

\item{num.trees}{number of trees for transformed outcome forest}

\item{mtry}{number of variables to possibly split at in each node}

\item{min.node.size}{minimum node size for transformed outcome forest}

\item{postprocess}{logical: should optional post-processing random forest be
fit at end?}

\item{verbose}{logical: should progress be printed to console?}
}
\value{
an object of class \code{PTOforest} with attributes:
 \itemize{
   \item x: matrix of covariates supplied by function call
   \item pscore: vector of propensity score supplied by function call
   \item postprocess: logical supplied by function call
   \item TOfit: fitted random forest on transformed outcomes
   \item PTOfit1: TOfit pollinated with treatment-arm outcomes
   \item PTOfit0: TOfit pollinated with control-arm outcomes
   \item postfit: post-processing random forest summarizing results
 }
}
\description{
Fit a pollinated transformed outcome (PTO) forest model
}
\examples{
# Randomized experiment example

n = 100 # number of training-set patients to simulate
p = 10  # number of features for each training-set patient

# Simulate data
x = matrix(rnorm(n * p), nrow = n, ncol = p) # simulate covariate matrix
tx_effect = x[, 1] + (x[, 2] > 0) # simple heterogeneous treatment effect
tx = rbinom(n, size = 1, p = 0.5) # random treatment assignment
y = rowMeans(x) + tx * tx_effect + rnorm(n, sd = 0.001) # simulate response

# Estimate PTO forest model
fit_pto = PTOforest(x, tx, y)
pred_pto = predict(fit_pto, newx = x)

# Visualize results
plot(tx_effect, pred_pto, main = 'PTO forest',
 xlab = 'True treatment effect', ylab = 'Estimated treatment effect')
abline(0, 1, lty = 2)

}
