% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Get propensity strata from propensity scores}
\usage{
stratify(pscore, tx, min.per.arm = 30)
}
\arguments{
\item{pscore}{vector of propensity scores}

\item{tx}{vector of treatment indicators}

\item{min.per.arm}{minimum number of observations for each arm within each
stratum}
}
\value{
a vector of integers with length equal to the length of pscore,
 reporting the propensity stratum corresponding to each propensity score
}
\description{
Get propensity strata from propensity scores
}
