% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagged.causalMARS.R
\name{bagged.causalMARS}
\alias{bagged.causalMARS}
\title{Fit a bag of causal MARS models}
\usage{
bagged.causalMARS(x, tx, y, nbag = 20, maxterms = 11, nquant = 5,
  degree = ncol(x), eps = 1, backstep = FALSE, propensity = FALSE,
  stratum = rep(1, nrow(x)), minnum = 5, verbose = FALSE)
}
\arguments{
\item{x}{matrix of covariates}

\item{tx}{vector of treatment indicators (0 or 1)}

\item{y}{vector of response values}

\item{nbag}{number of models to bag}

\item{maxterms}{maximum number of terms to include in the regression basis
(e.g. \code{maxterms = 11} means intercept + 5 pairs added)}

\item{nquant}{number of quantiles used in splitting}

\item{degree}{max number of different predictors that can interact in model}

\item{eps}{shrinkage factor for new term added}

\item{backstep}{logical: should out-of-bag samples be used to prune each
model? otherwise full regression basis is used for each model}

\item{propensity}{logical: should propensity score stratification be used?}

\item{stratum}{optional vector giving propensity score stratum for each
observation (only used if \code{propensity = TRUE})}

\item{minnum}{minimum number of observations in each arm of each propensity
score stratum needed to estimate regression coefficients for basis
(only used if \code{propensity = TRUE})}

\item{verbose}{logical: should progress be printed to console?}
}
\value{
an object of class \code{bagged.causalMARS}, which is itself a list
 of \code{causalMARS} objects
}
\description{
Fit a bag of causal MARS models
}
